Dedupe `e_elevator`
