+++ config/splat.pspeu.sttop.yaml
@@ -61,7 +61,8 @@ segments:
       - [0x10BE8, c, top_psp/st_init]
       - [0x10FB8, c, top/e_breakable]
       - [0x118F0, c, top/post_richter_cutscene]
-      - [0x124B8, c, top_psp/unk_12048]
+      - [0x124B8, c, top/e_elevator]
+      - [0x137D0, c, top_psp/unk_137D0]
       - [0x14930, c, top_psp/cutscene]
       - [0x17738, c, top/e_room_bg]
       - [0x17818, c, top/e_lock_camera]
@@ -88,7 +89,8 @@ segments:
       - [0x20BC0, data]
       - [0x3E6E0, .data, top/header]
       - [0x3E718, .data, top/post_richter_cutscene]
-      - [0x3E758, data]
+      - [0x3E758, .data, top/e_elevator]
+      - [0x3E780, data]
       - [0x3EC70, .data, top/e_room_bg]
       - [0x3ECB0, .data, top/e_lock_camera]
       - [0x3ED30, .data, top/e_richter_fight_trigger]
@@ -99,7 +101,7 @@ segments:
       - [0x3F2A8, .rodata, top/e_red_door]
       - [0x3F2C0, .rodata, top/e_collect]
       - [0x3F320, .rodata, top/e_flea_rider]
-      - [0x3F348, .rodata, top_psp/unk_12048]
+      - [0x3F348, .rodata, top_psp/unk_137D0]
       - [0x3F380, .rodata, top_psp/cutscene]
       - [0x3F480, .bss, top/create_entity]
       - [0x3F4A0, .bss, top/st_update]
