+++ config/splat.us.sttop.yaml
@@ -46,6 +46,8 @@ segments:
       - [0xA98, .data, e_lock_camera]
       - [0xAF8, .data, e_breakable]
       - [0xC08, data, 296F4]
+      - [0xCC8, .data, e_elevator]
+      - [0xCF0, data]
       - [0xD10, .data, d_prize_drops]
       - [0xD38, data]
       - [0xDF0, .data, post_richter_cutscene]
@@ -67,6 +69,7 @@ segments:
       - [0x28904, .rodata, e_richter_fight_trigger]
       - [0x2891C, .rodata, e_breakable]
       - [0x2893C, .rodata, 296F4]
+      - [0x28978, .rodata, 2B308]
       - [0x2898C, .rodata, cutscene]
       - [0x28A30, .rodata, post_richter_cutscene]
       - [0x28A64, .rodata, e_red_door]
@@ -80,6 +83,8 @@ segments:
       - [0x29250, c, e_richter_fight_trigger]
       - [0x29390, c, e_breakable]
       - [0x299EC, c, 296F4]
+      - [0x2A624, c, e_elevator]
+      - [0x2B308, c, 2B308]
       - [0x2B768, c, cutscene]
       - [0x2D11C, c, post_richter_cutscene]
       - [0x2D630, c, st_update]
