+++ config/symbols.pspeu.sttop.txt
@@ -29,7 +29,6 @@ AllocEntity = 0x09246B60;
 SetStep = 0x09247070;
 InitializeEntity = 0x09247190;
 EntityDummy = 0x09247338;
-GetPlayerCollisionWith = 0x09247500;
 GetLangAt = 0x09248268;
 InitEntityIds = 0x09248330;
 TOP_Load = 0x09248480;
@@ -38,6 +37,8 @@ EntityBreakableDebris = 0x09248B90;
 func_us_801AD11C = 0x09248F70;
 func_us_801AD320 = 0x09249208;
 func_us_801AD400 = 0x09249388;
+EntityTriangleElevator = 0x09249B38;
+func_us_801AABA4 = 0x0924A3B0;
 EntitySecretButton = 0x0924AE50;
 EntitySecretStairs = 0x0924B4F0;
 EntityRedEyeBust = 0x0924BB48;
@@ -71,6 +72,7 @@ g_EInitUnkId13 = 0x09257088;
 g_EInitLockCamera = 0x09257098;
 g_EInitCommon = 0x092570A8;
 g_EInitDamageNum = 0x092570B8;
+g_EInitTOPCommon = 0x092570C8;
 g_EInitFleaRider = 0x092570D8;
 g_EInitAxeKnight = 0x092570E8;
 g_EInitAxeKnightAxe = 0x092570F8;
@@ -79,6 +81,7 @@ D_us_80182E44 = 0x092580F0;
 TOP_Overlay = 0x09275D60;
 TOP_PrizeDrops = 0x09275E00;
 D_psp_E_BREAKABLE_DEBRIS = 0x09276BD8;
+D_us_801BC514 = 0x09276C90;
 g_Dialogue = 0x09277258; // size:0x40
 g_SkipCutscene = 0x09277298;
 g_CutsceneFlags = 0x092772A0;
