+++ config/symbols.us.sttop.txt
@@ -5,7 +5,6 @@ rooms_layers = 0x801805C8;
 gfxBanks = 0x801806F0;
 entityLayoutHorizontal = 0x80180740;
 entityLayoutVertical = 0x80180814;
-D_us_80180CC8 = 0x80180CC8;
 LionLampAnim = 0x80180CF0;
 actor_names = 0x80180DE4;
 rooms = 0x80181B78;
